+++ agave/frames/ethernet.py
@@ -1,3 +1,4 @@
+from typing import Union
 from . core import Frame
 
 
@@ -6,12 +7,34 @@ ETHER_TYPE_IPV4 = 0x0800
 ETHER_TYPE_IPV6 = 0x86dd
 
 
+class MACAddress:
+
+	__slots__ = ["address"]
+
+	def __init__(self, address: Union[bytes, str]):
+		if type(address) == str:
+			self.address = self.str_to_mac(address)
+		else:
+			self.address = address
+
+	@classmethod
+	def str_to_mac(cls, address: str):
+		return bytes(map(lambda x: int(x, 16), address.split(":")))
+
+	@classmethod
+	def mac_to_str(cls, address: bytes):
+		return ':'.join('%02x'%i for i in address)
+
+	def __str__(self):
+		return self.mac_to_str(self.address)
+
+
 def mac_to_str(address: bytes) -> str:
-	return ':'.join('%02x'%i for i in address)
+	return MACAddress.mac_to_str(address)
 
 
 def str_to_mac(address: str) -> bytes:
-	return bytes(map(lambda x: int(x, 16), address.split(":")))
+	return MACAddress.str_to_mac(address)
 
 
 class Ethernet(Frame):
